considering big numbers

+++ src/main/java/taumandbirthday/DikshaGifts.java
@@ -12,7 +12,7 @@ public class DikshaGifts {
                     scanner.nextInt(), scanner.nextInt(), scanner.nextInt()));
     }
 
-    public static int calculate(int b, int w, int x, int y, int z) {
+    public static long calculate(long b, long w, long x, long y, long z) {
         return  b * (x <= y + z ? x : y + z ) +
                 w * (y <= x + z ? y : x + z);
     }
